Explicit cast as required by the NEC compiler.
Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@1442 17b73243-c579-4c4c-a9d2-2d5706c11dac

+++ src/util/Cache.c
@@ -91,7 +91,7 @@ void *Util_CacheMalloc(unsigned index,
     cache_size = 0;
   }
 
-  data = (char *)malloc(size+cache_size);
+  data = (char *)malloc((size_t)(size+cache_size));
 
   if(data)
   {
